Frontend/requirements.txt
fastapi[all]
google-cloud-storage